requirements.txt
Flask
Flask-SQLAlchemy
Flask-Session
Flask-WTF
Flask-Login
Flask-Flash
Flask-Bcrypt
pymysql
requests
sqlite3
qrcode
Pillow
datetime
platformdirs
os
Werkzeug
gunicorn
waitress
